Cleaned up a lot of issues with anvil e2e tests #539
