+++ explorer/.gitignore
@@ -7,6 +7,7 @@
 
 # testing
 /coverage
+/test-results/*
 
 # next.js
 /.next/
